requirements.txt
aiomysql
aiomcache
aiohttp
asyncio
PyYAML
jinja2
aiohttp-jinja2
multidict